Move convert_code_blocks to conf
